refactor(agent_provider): ClaudeTransport enum + dispatcher (#748)
Split src/agent_provider/claude.rs into a directory module:
- claude/mod.rs   — ClaudeProvider, ClaudeTransport, InteractiveDriver,
                    and a transport dispatcher in AgentProvider::run.
- claude/headless.rs — the existing one-shot subprocess logic, relocated
                    verbatim (build_stream_args/build_text_args/run_text/
                    health_check/run). Behaviour preserved bit-for-bit.

ClaudeProvider::new and Default set transport: Headless, so every existing
call path is unchanged. The Interactive arm returns AgentError::Config
("interactive transport not implemented") without spawning — the seam where
the PTY backend (#749) plugs in next.

The four pub methods (build_stream_args, build_text_args, run_text,
health_check_blocking) stay inherent on ClaudeProvider and delegate to
pub(super) free fns in headless.rs, preserving the public surface used by
manager.rs, doctor.rs, adapt/prompts.rs, sanitize/analyzer.rs, and
review/auto_review.rs.

AC deviation: the issue specified `enum InteractiveDriver {}` (uninhabited),
but that makes the dispatch arm dead code and its required test impossible to
construct. Replaced with a single `Unimplemented` marker variant (minimal
inhabited form), tagged `// TODO(#749)`.

Cites RUST-GUARDRAILS.md §1 (module discipline) and §5 (subprocess lifecycle
preserved). Each new file <= 400 LOC.
